+++ v2/backend/canisters/user/impl/src/model/group_chats.rs
@@ -5,11 +5,11 @@ use std::collections::hash_map::Entry::{Occupied, Vacant};
 use std::collections::HashMap;
 use types::{ChatId, TimestampMillis};
 
-const MAX_GROUPS_PER_USER: u16 = 10;
+const MAX_GROUPS_PER_USER: u32 = 10;
 
 #[derive(CandidType, Deserialize, Default)]
 pub struct GroupChats {
-    groups_created: u16,
+    groups_created: u32,
     group_chats: HashMap<ChatId, GroupChat>,
 }
 
@@ -56,7 +56,7 @@ impl GroupChats {
         self.group_chats.values()
     }
 
-    pub fn max_groups_created(&self) -> Option<u16> {
+    pub fn max_groups_created(&self) -> Option<u32> {
         if self.groups_created >= MAX_GROUPS_PER_USER {
             Some(MAX_GROUPS_PER_USER)
         } else {
